refactor: unit coordinate 관련 패키지 분리
