+++ src/main/java/coordinates/domain/point/unitcoordinate/UnitCoordinatePool.java
@@ -1,4 +1,4 @@
-package coordinates.domain.point;
+package coordinates.domain.point.unitcoordinate;
 
 import java.util.HashMap;
 import java.util.Map;
