+++ src/test/java/coordinates/domain/point/unitcoordinate/RelativeCoordinatePoolTest.java
@@ -1,4 +1,4 @@
-package coordinates.domain.point;
+package coordinates.domain.point.unitcoordinate;
 
 import org.junit.jupiter.api.Test;
