+++ src/test/java/coordinates/domain/point/unitcoordinate/XCoordinateTest.java
@@ -1,4 +1,4 @@
-package coordinates.domain.point;
+package coordinates.domain.point.unitcoordinate;
 
 import coordinates.exception.OutOfXCoordinateBoundException;
 import org.junit.jupiter.api.BeforeEach;
