+++ src/test/java/coordinates/domain/point/unitcoordinate/YCoordinateTest.java
@@ -1,4 +1,4 @@
-package coordinates.domain.point;
+package coordinates.domain.point.unitcoordinate;
 
 import coordinates.exception.OutOfYCoordinateBoundException;
 import org.junit.jupiter.api.BeforeEach;
